+++ restful-service/src/main/java/vn/edu/fpt/sba/dto/response/AlbumResponseDTO.java
@@ -1,6 +1,6 @@
 package vn.edu.fpt.sba.dto.response;
 
 public record AlbumResponseDTO(
-        Long albumId,
+        Integer albumId,
         String title) {
 }
